+++ rqalpha/__init__.py
@@ -19,7 +19,9 @@ RQAlpha - a Algorithm Trading System
 """
 
 import pkgutil
+from .__main__ import cli
 from .cmd import cmd_cli
+from .api.api_base import export_as_api
 
 __all__ = [
     '__version__',