+++ rqalpha/api/api_base.py
@@ -27,6 +27,7 @@ import sys
 from collections import Iterable
 from functools import wraps
 from types import FunctionType
+# noinspection PyUnresolvedReferences
 from typing import List
 
 import pandas as pd
@@ -39,6 +40,7 @@ from ..execution_context import ExecutionContext
 from ..utils import to_industry_code, to_sector_name, unwrapper
 from ..utils.exception import patch_user_exc, patch_system_exc, EXC_EXT_NAME, RQInvalidArgument
 from ..utils.i18n import gettext as _
+# noinspection PyUnresolvedReferences
 from ..utils.logger import user_log as logger
 
 from ..model.instrument import SectorCodeItem, IndustryCodeItem
@@ -118,6 +120,7 @@ def export_as_api(func):
     __all__.append(func.__name__)
 
     func = decorate_api_exc(func)
+    globals()[func.__name__] = func
 
     return func
 
@@ -327,9 +330,10 @@ def get_yield_curve(date=None, tenor=None):
              verify_that('frequency').is_valid_frequency(),
              verify_that('fields').are_valid_fields(names.VALID_HISTORY_FIELDS, ignore_none=True),
              verify_that('skip_suspended').is_instance_of(bool),
+             verify_that('include_now').is_instance_of(bool),
              verify_that('adjust_type').is_in({'pre', 'none', 'post'}))
 def history_bars(order_book_id, bar_count, frequency, fields=None, skip_suspended=True,
-                 adjust_type='pre'):
+                 include_now=False, adjust_type='pre'):
     """
     获取指定合约的历史行情，同时支持日以及分钟历史数据。不能在init中调用。 注意，该API会自动跳过停牌数据。
 
@@ -386,6 +390,8 @@ def history_bars(order_book_id, bar_count, frequency, fields=None, skip_suspende
     prev_settlement             结算价（期货日线专用）
     =========================   ===================================================
 
+    :param bool skip_suspended: 是否跳过停牌数据
+    :param bool include_now: 是否包含当前数据
     :param str adjust_type: 复权类型，默认为前复权 pre；可选 pre, none, post
 
     :return: `ndarray`, 方便直接与talib等计算库对接，效率较history返回的DataFrame更高。
@@ -404,7 +410,6 @@ def history_bars(order_book_id, bar_count, frequency, fields=None, skip_suspende
     """
     order_book_id = assure_order_book_id(order_book_id)
     env = Environment.get_instance()
-
     dt = env.calendar_dt
 
     if frequency[-1] == 'm' and env.config.base.frequency == '1d':
@@ -413,13 +418,16 @@ def history_bars(order_book_id, bar_count, frequency, fields=None, skip_suspende
     if adjust_type not in {'pre', 'post', 'none'}:
         raise RuntimeError('invalid adjust_type')
 
-    if (env.config.base.frequency == '1m' and frequency == '1d') or \
-        (frequency == '1d' and ExecutionContext.phase == EXECUTION_PHASE.BEFORE_TRADING):
-        # 在分钟回测获取日线数据, 应该推前一天，这里应该使用 trading date
-        dt = env.data_proxy.get_previous_trading_date(env.trading_dt)
+    if frequency == '1d':
+        sys_frequency = Environment.get_instance().config.base.frequency
+        if ((sys_frequency == '1m' and not include_now) or
+                ExecutionContext.phase == EXECUTION_PHASE.BEFORE_TRADING):
+            dt = env.data_proxy.get_previous_trading_date(env.trading_dt)
+            include_now = False
 
     return env.data_proxy.history_bars(order_book_id, bar_count, frequency, fields, dt,
-                                       skip_suspended, adjust_type, env.trading_dt)
+                                       skip_suspended=skip_suspended, include_now=include_now,
+                                       adjust_type=adjust_type, adjust_orig=env.trading_dt)
 
 
 @export_as_api