+++ rqalpha/examples/data_source/import_get_csv_module.py
@@ -1,0 +1,33 @@
+from rqalpha.api import *
+import os
+import sys
+
+
+def init(context):
+    strategy_file_path = context.config.base.strategy_file
+    sys.path.append(os.path.realpath(os.path.dirname(strategy_file_path)))
+
+    from get_csv_module import get_csv
+
+    IF1706_df = get_csv()
+    context.IF1706_df = IF1706_df
+
+
+def before_trading(context):
+    logger.info(context.IF1706_df)
+
+
+__config__ = {
+    "base": {
+        "securities": "future",
+        "start_date": "2015-01-09",
+        "end_date": "2015-01-10",
+        "frequency": "1d",
+        "matching_type": "current_bar",
+        "future_starting_cash": 1000000,
+        "benchmark": None,
+    },
+    "extra": {
+        "log_level": "verbose",
+    },
+}