+++ rqalpha/examples/macd.py
@@ -26,7 +26,7 @@ def handle_bar(context, bar_dict):
     # TODO: 开始编写你的算法吧！
 
     # 读取历史数据，使用sma方式计算均线准确度和数据长度无关，但是在使用ema方式计算均线时建议将历史数据窗口适当放大，结果会更加准确
-    prices = history_bars(context.s1, context.OBSERVATION,'1d','close')
+    prices = history_bars(context.s1, context.OBSERVATION, '1d', 'close')
 
     # 用Talib计算MACD取值，得到三个时间序列数组，分别为macd, signal 和 hist
     macd, signal, hist = talib.MACD(prices, context.SHORTPERIOD,
@@ -42,7 +42,7 @@ def handle_bar(context, bar_dict):
     if macd[-1] - signal[-1] < 0 and macd[-2] - signal[-2] > 0:
         # 计算现在portfolio中股票的仓位
         curPosition = context.portfolio.positions[context.s1].quantity
-        #进行清仓
+        # 进行清仓
         if curPosition > 0:
             order_target_value(context.s1, 0)