+++ rqalpha/model/account/stock_account.py
@@ -13,9 +13,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from collections import defaultdict
 
 import six
+from collections import defaultdict
 
 from .base_account import BaseAccount
 from ...events import EVENT
@@ -104,8 +104,9 @@ class StockAccount(BaseAccount):
         if event.account != self:
             return
         order = event.order
-        position = self._positions.get_or_create(order.order_book_id)
-        position.on_order_pending_new_(order)
+        position = self._positions.get(order.order_book_id, None)
+        if position is not None:
+            position.on_order_pending_new_(order)
         if order.side == SIDE.BUY:
             order_value = order.frozen_price * order.quantity
             self._frozen_cash += order_value