+++ rqalpha/utils/default_future_info.py
@@ -15,8 +15,9 @@
 # limitations under the License.
 
 from datetime import time
+
 from .datetime_func import TimeRange
-from ..const import COMMISSION_TYPE, MARGIN_TYPE
+
 
 NIGHT_UNDERLYING_SYMBOL = ["CU", "AL", "ZN", "PB", "SN", "NI", "RB", "HC", "BU", "RU", "AU", "AG", "Y", "M", "A", "B",
                            "P", "J", "JM", "I", "CF", "SR", "OI", "MA", "ZC", "FG", "RM"]