+++ rqalpha/utils/rq_json.py
@@ -16,6 +16,7 @@
 
 import simplejson as json
 import datetime
+
 from .. import const