+++ CHANGELOG.rst
@@ -2,10 +2,26 @@
 CHANGELOG
 ==================
 
+2.1.0
+==================
+
+- Fix `Issue 87 <https://github.com/ricequant/rqalpha/issues/87>`_
+- Fix `Issue 89 <https://github.com/ricequant/rqalpha/pull/89>`_
+- Fix 无法通过 :code:`env.config.mod` 获取全部 `mod` 的配置信息
+- 增加 :code:`context.config` 来获取配置信息
+- 提供 :code:`from rqalpha import export_as_api` 接口，方便扩展自定义 API
+
+2.0.9
+==================
+
+- Fix `Issue 79 <https://github.com/ricequant/rqalpha/issues/79>`_
+- Fix `Issue 82 <https://github.com/ricequant/rqalpha/issues/82>`_
+- Fix :code:`rqalpha cmd` 失效
+
 2.0.8
 ==================
 
-- Fix [Issue 81](https://github.com/ricequant/rqalpha/issues/81)
+- Fix `Issue 81 <https://github.com/ricequant/rqalpha/issues/81>`_
 - 解决 `mod_config.yml` 文件解析出错以后，所有的命令报错的问题
 - 默认在 Python 2.x 下 `sys.setdefaultencoding("utf-8")`
 - 优化 `UNIVERSE_CHANGED` 事件，现在只有在universe真正变化时才触发
@@ -13,7 +29,7 @@ CHANGELOG
 2.0.7
 ==================
 
-- Fix [Issue 78](https://github.com/ricequant/rqalpha/issues/78)
+- Fix `Issue 78 <https://github.com/ricequant/rqalpha/issues/78>`_
 - `is_st_stock` | `is_suspended` 支持 `count` 参数
 - 解决大量 Python 2.x 下中文乱码问题